+++ lib/challenge_gov/reports.ex
@@ -106,16 +106,4 @@ defmodule ChallengeGov.Reports do
         {datetime_start, datetime_end}
     end
   end
-
-  def generate_date_options do
-    months =
-      Enum.reduce(1..12, [], fn num, acc ->
-        Enum.concat(acc, [{Timex.month_name(num), num}])
-      end)
-
-    days = Range.new(1, 31)
-    years = Range.new(Timex.now().year, 2020)
-
-    [years, months, days]
-  end
 end
