+++ lib/web/templates/reports/index.html.eex
@@ -1,0 +1,63 @@
+<div class="content-header">
+  <div class="container-fluid">
+    <%= SharedView.render_breadcrumbs([
+      %{text: "Home", route: Routes.dashboard_path(@conn, :index)},
+      %{text: "Reports"},
+    ])%>
+    <div class="row mb-2">
+      <div class="col-sm-6">
+        <h1 class="m-0 text-dark">
+          Reports
+        </h1>
+      </div>
+    </div>
+  </div>
+</div>
+
+<section class="content">
+  <div class="row">
+    <div class="col-md-12 row content-sub-1">
+      <h4 class="box-title">Security Log</h4>
+      <%= link(to: Routes.reports_path(@conn, :export_security_log), class: "btn component-btn mx-3") do %>
+        <span>Export All</span>
+      <% end %>
+    </div>
+    <div class="col-md-3 content-sub-1">
+      <h5>Filter by date:</h5>
+      <div class="col-md-6 p-0">
+        <%= form_for(@changeset, Routes.reports_path(@conn, :export_security_log), [method: :get], fn f -> %>
+          <%= FormView.select_field(f, :year, collection: Enum.map(@years, &(&1)), prompt: "", label: "Year", class: "form-control", required: true) %>
+          <%= FormView.select_field(f, :month, collection: Enum.map(@months, &(&1)), prompt: "", label: "Month", class: "form-control") %>
+          <%= FormView.select_field(f, :day, collection: Enum.map(@days, &(&1)), prompt: "", label: "Day", class: "form-control") %>
+          <div class="ch-export-btn-wrapper">
+            <%= submit("Export", class: "btn component-btn report-export") %>
+          </div>
+        <% end) %>
+      </div>
+    </div>
+  </div>
+</section>
+
+<hr class="ch-report-hr"/>
+
+<section class="content">
+  <div class="row">
+    <div class="col-md-12 row content-sub-1">
+      <h4 class="box-title">Certification Log</h4>
+      <%= link(to: Routes.reports_path(@conn, :export_certification_log), class: "btn component-btn mx-3") do %>
+        <span>Export All</span>
+      <% end %>
+    </div>
+    <div class="col-md-3 content-sub-1">
+      <h5>Filter by year certified:</h5>
+      <div class="col-md-6 p-0">
+        <%= form_for(@conn, Routes.reports_path(@conn, :export_certification_log), [method: :get], fn f -> %>
+          <%= FormView.select_field(f, :year, collection: Enum.map(@years, &(&1)), prompt: "", label: "Year", class: "form-control", required: true) %>
+          <div class="ch-export-btn-wrapper">
+            <%= submit("Export", class: "btn component-btn") %>
+          </div>
+        <% end) %>
+      </div>
+    </div>
+  </div>
+</section>
